Degrain: fix integer overflow with 16 bit input
With large enough thsad, the filter would become a very slow no-op.

+++ src/MVDegrains.cpp
@@ -681,9 +681,8 @@ static void VS_CC mvdegrainCreate(const VSMap *in, VSMap *out, void *userData, V
         return;
     }
 
-    d.thSAD[0] = d.thSAD[0] * d.mvClips[Backward1]->GetThSCD1() / d.nSCD1; // normalize to block SAD
-    d.thSAD[1] = d.thSAD[2] = d.thSAD[1] * d.mvClips[Backward1]->GetThSCD1() / d.nSCD1; // chroma threshold, normalized to block SAD
-
+    d.thSAD[0] = (int64_t)d.thSAD[0] * d.mvClips[Backward1]->GetThSCD1() / d.nSCD1; // normalize to block SAD
+    d.thSAD[1] = d.thSAD[2] = (int64_t)d.thSAD[1] * d.mvClips[Backward1]->GetThSCD1() / d.nSCD1; // chroma threshold, normalized to block SAD
 
 
     d.node = vsapi->propGetNode(in, "clip", 0, 0);
